+++ src/com/intellij/advancedExpressionFolding/processor/controlflow/PsiTryStatementExt.kt
@@ -4,7 +4,7 @@ import com.intellij.advancedExpressionFolding.expression.Expression
 import com.intellij.advancedExpressionFolding.expression.semantic.HideExpression
 import com.intellij.advancedExpressionFolding.processor.*
 import com.intellij.advancedExpressionFolding.processor.core.BaseExtension
-import com.intellij.advancedExpressionFolding.processor.lombok.MethodBodyInspector.getRethrownException
+import com.intellij.advancedExpressionFolding.processor.lombok.MethodBodyInspector.rethrownException
 import com.intellij.psi.PsiStatement
 import com.intellij.psi.PsiTryStatement
 
@@ -20,7 +20,7 @@ object PsiTryStatementExt : BaseExtension() {
             it.catchBlock?.statements?.singleOrNull() != null
         } ?: return null
 
-        val rethrownException = catchBlock.getRethrownException() ?: return null
+        val rethrownException = catchBlock.rethrownException ?: return null
 
         val tryKeyword = firstChild
         val sneakyThrows = when {