+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+- Add compatibility with IntelliJ IDEA 2025.3 EAP (253.20558.43)
+
 ## [4.2.0] - 2025-09-20
 
 - [Add setting to prevent collapsing Java text blocks in log folding](https://github.com/AntoniRokitnicki/AdvancedExpressionFolding/pull/338)